fixed nav: added link to tags view; make anchors qualified
This patch changes the default layout template ('wrapper.tt') to make
the nav bar at the top contain a link to the /tags/ view.  It also
changes the links to 'recent.html' and 'atom.xml' to be anchor at the
docroot's base.  This is needed so that the nav works from inside the
/tags/ view.